requirements.txt
# 问卷转换器Web应用依赖包

# Web框架
Flask==2.3.3
Werkzeug==2.3.7

# 文档处理
python-docx==0.8.11

# 数据处理和XML处理
lxml==4.9.3

# WSGI服务器 (生产环境)
gunicorn==21.2.0

# 跨域支持 (可选)
Flask-CORS==4.0.0

# 环境配置
python-dotenv==1.0.0

# 基础依赖 (Flask自动包含，但明确列出)
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7

# 文件监控
pathlib2==2.3.7

# 编码检测
chardet==5.2.0

# 临时文件处理
tempfile2==0.1.0